@@ -11,7 +11,9 @@
     "db-seed": "node prisma/seed.mjs",
     "deploy": "npm run build && npm run db-migrate-deploy && npm run db-seed",
     "lint": "eslint .",
-    "lint-fix": "eslint . --fix"
+    "lint-fix": "eslint . --fix",
+    "tsc": "./node_modules/typescript/bin/tsc",
+    "check": "eslint . && npm run tsc"
   },
   "dependencies": {
     "@geist-ui/react": "^2.1.3",
@@ -19,7 +21,7 @@
     "@onflow/fcl": "0.0.70",
     "@onflow/types": "0.0.4",
     "@onflow/util-encode-key": "0.0.2",
-    "@prisma/client": "^2.17.0",
+    "@prisma/client": "^2.24.1",
     "date-fns": "^2.11.1",
     "elliptic": "^6.5.4",
     "formik": "^2.2.6",
@@ -36,13 +38,22 @@
     "yup": "^0.32.9"
   },
   "devDependencies": {
+    "@types/elliptic": "^6.4.12",
+    "@types/hcaptcha__react-hcaptcha": "^0.1.4",
+    "@types/mixpanel-browser": "^2.35.6",
+    "@types/node": "^15.12.2",
+    "@types/react": "^17.0.11",
+    "@typescript-eslint/eslint-plugin": "^4.26.1",
+    "@typescript-eslint/parser": "^4.26.1",
     "babel-eslint": "^10.1.0",
     "dotenv": "^8.2.0",
-    "eslint": "^7.21.0",
+    "eslint": "^7.28.0",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-prettier": "^3.3.1",
     "eslint-plugin-react": "^7.22.0",
+    "eslint-plugin-react-hooks": "^4.2.0",
     "prettier": "^2.2.1",
-    "prisma": "^2.17.0"
+    "prisma": "^2.24.1",
+    "typescript": "^4.3.2"
   }
 }
